refactor(audit): split discovery support and tests
